drunk_fov gameplay mod

+++ cl_dll/hud.h
@@ -881,7 +881,7 @@ public:
 	Vector	m_vecAngles;
 	int		m_iKeyBits;
 	int		m_iHideHUDDisplay;
-	int		m_iFOV;
+	float	m_iFOV;
 	int		m_Teamplay;
 	int		m_iRes;
 	cvar_t  *m_pCvarStealMouse;
